test(valecup): split the suite along describe boundaries for shard balance
tests/vale_cup.test.ts was the suite's single heaviest file and vitest's
sha1-based shard assignment cannot subdivide a file, so the worst shard of
the Phase 4 matrix always carried its whole cost. Split it into three
suites along existing describe boundaries (a pure move: test bodies are
byte-identical, 64 tests before and after):

- tests/vale_cup_bots.test.ts: bot showcase, bot backfill and practice
- tests/vale_cup_match.test.ts: possession gate, queue guards, matchmaking
  and packing, match lifecycle, sport moves, desertion, the closed pitch
- tests/vale_cup_meta.test.ts: parimutuel betting, kit swap round trip and
  persistence, determinism, parallel private practice, guild banners

The shared world-staging helpers (makeWorld, teleport, addAt, errorsOf,
readyAll, startBout, tickUntil) moved verbatim to tests/vale_cup_util.ts
(not a test file; imported by the three suites).

Measured on the sha1 shard simulation (vitest 4.1.8 BaseSequencer): the
three names land in three different quarters of an N=4 matrix, flattening
the worst shard from 252.5s to 230.0s aggregate local cost.
